Aplicando Boas práticas de implementação

+++ ApiFuncional/Controllers/ProdutosController.cs
@@ -1,7 +1,9 @@
+using System.Data;
 using ApiFuncional.Data;
 using ApiFuncional.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 
 namespace ApiFuncional.Controllers
 {
@@ -16,22 +18,39 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
         {
+            if (_context.Produtos == null) return NotFound();
+
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+            if (_context.Produtos == null) return NotFound();
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null) return NotFound();
 
             return produto;
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
+            if (_context.Produtos == null) return Problem("Erro ao criar um produto, contate o suporte!");
+
             if (!ModelState.IsValid)
             {
                 //return BadRequest(ModelState); //retorna somente o error dos models
@@ -49,27 +68,59 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpPut("{id}:int")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> PutProduto(int id, Produto produto)
         {
             if (id != produto.Id) return BadRequest();
 
             if (!ModelState.IsValid) return ValidationProblem(ModelState);
+
+            _context.Entry(produto).State = EntityState.Modified; //nn precisa atachar, só modifica.
+            // _context.Produtos.Update(produto); //atacha primeiro, se o mesmo ja foi manipulado anteriormente, nesse formato, geraria um erro.
+
+            //tratando concorrência
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DBConcurrencyException)
+            {
+                if (!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return NoContent();
+        }
+
+        [HttpDelete("{id}:int")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
+        public async Task<IActionResult> DeleteProduto(int id)
+        {
+            if (_context.Produtos == null) return NotFound();
+
+            var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null) return NotFound();
             
-            _context.Produtos.Update(produto);
+            _context.Produtos.Remove(produto);
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
 
-        [HttpDelete("{id}:int")]
-        public async Task<IActionResult> PutProduto(int id)
+        private bool ProdutoExists(int id)
         {
-            var produto = await _context.Produtos.FindAsync(id);
-
-            _context.Produtos.Remove(produto);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
